requirements.txt
openai>=1.0.0
pandas>=2.0.0
streamlit>=1.28.0
plotly>=5.17.0
scipy>=1.11.0
python-dotenv>=1.0.0